Refactor CLI options and arguments to make more sense
Instead of multiple input file handling, the application is simplified
to just a single input template and single output result with a mapping
required for the only argument. The `-i,--input` and `-o,--output`
options are added to change the source stream from stdin and the
destination stream from stdout.