Add build script
The build script will build the manpage from the markdown source using
pandoc if the release profile is used, i.e. `cargo build --release`. The
output is in the `man` folder as `groom.1`. This is a small workaround
until Cargo adds support for manpages. It looks like manpages will
eventually be recognized and included in the `cargo install` command if
placed in a `man` folder adjacent to the manifest file.